Don't load ActionView when building the middleware stack
