+++ actionpack/lib/action_dispatch/middleware/debug_exceptions.rb
@@ -4,9 +4,6 @@ require "action_dispatch/http/request"
 require "action_dispatch/middleware/exception_wrapper"
 require "action_dispatch/routing/inspector"
 require "action_view"
-require "action_view/base"
-
-require "pp"
 
 module ActionDispatch
   # This middleware is responsible for logging exceptions and
@@ -14,41 +11,7 @@ module ActionDispatch
   class DebugExceptions
     RESCUES_TEMPLATE_PATH = File.expand_path("templates", __dir__)
 
-    class DebugView < ActionView::Base
-      def debug_params(params)
-        clean_params = params.clone
-        clean_params.delete("action")
-        clean_params.delete("controller")
-
-        if clean_params.empty?
-          "None"
-        else
-          PP.pp(clean_params, +"", 200)
-        end
-      end
-
-      def debug_headers(headers)
-        if headers.present?
-          headers.inspect.gsub(",", ",\n")
-        else
-          "None"
-        end
-      end
-
-      def debug_hash(object)
-        object.to_hash.sort_by { |k, _| k.to_s }.map { |k, v| "#{k}: #{v.inspect rescue $!.message}" }.join("\n")
-      end
-
-      def render(*)
-        logger = ActionView::Base.logger
-
-        if logger && logger.respond_to?(:silence)
-          logger.silence { super }
-        else
-          super
-        end
-      end
-    end
+    autoload :DebugView, "action_dispatch/middleware/debug_exceptions/debug_view"
 
     cattr_reader :interceptors, instance_accessor: false, default: []
 
@@ -200,7 +163,7 @@ module ActionDispatch
       end
 
       def logger(request)
-        request.logger || ActionView::Base.logger || stderr_logger
+        request.logger || DebugView.logger || stderr_logger
       end
 
       def stderr_logger
